+++ lioamber/fileio/restart_rho.f90
@@ -29,7 +29,7 @@ subroutine read_rho_restart_ccd(rho, M, UID)
    ! UID : Input file unit ID.
    implicit none
    integer   , intent(in)  :: M, UID
-   complex*16, intent(out) :: rho(M,M)
+   complex(kind=8), intent(out) :: rho(M,M)
 
    rewind(UID)
    call read_sqmatrix(rho, M, UID)
@@ -42,7 +42,7 @@ subroutine read_rho_restart_ccs(rho, M, UID)
    ! UID : Input file unit ID.
    implicit none
    integer  , intent(in)  :: M, UID
-   complex*8, intent(out) :: rho(M,M)
+   complex(kind=4), intent(out) :: rho(M,M)
 
    rewind(UID)
    call read_sqmatrix(rho, M, UID)
@@ -82,7 +82,7 @@ subroutine read_rho_restart_ocd(rho_a, rho_b, M, UID)
    ! UID   : Input file unit ID.
    implicit none
    integer   , intent(in)  :: M, UID
-   complex*16, intent(out) :: rho_a(M,M), rho_b(M,M)
+   complex(kind=8), intent(out) :: rho_a(M,M), rho_b(M,M)
 
    rewind(UID)
    call read_sqmatrix(rho_a, M, UID)
@@ -97,7 +97,7 @@ subroutine read_rho_restart_ocs(rho_a, rho_b, M, UID)
    ! UID   : Input file unit ID.
    implicit none
    integer  , intent(in)  :: M, UID
-   complex*8, intent(out) :: rho_a(M,M), rho_b(M,M)
+   complex(kind=4), intent(out) :: rho_a(M,M), rho_b(M,M)
 
    rewind(UID)
    call read_sqmatrix(rho_a, M, UID)
@@ -147,7 +147,7 @@ subroutine write_rho_restart_ccd(rho, M, UID)
    ! UID : Output file unit ID.
    implicit none
    integer   , intent(in) :: M, UID
-   complex*16, intent(in) :: rho(M,M)
+   complex(kind=8), intent(in) :: rho(M,M)
 
    rewind(UID)
    call write_sqmatrix(rho, M, UID)
@@ -160,7 +160,7 @@ subroutine write_rho_restart_ccs(rho, M, UID)
    ! UID : Output file unit ID.
    implicit none
    integer  , intent(in) :: M, UID 
-   complex*8, intent(in) :: rho(M,M)
+   complex(kind=4), intent(in) :: rho(M,M)
 
    rewind(UID)
    call write_sqmatrix(rho, M, UID)
@@ -200,7 +200,7 @@ subroutine write_rho_restart_ocd(rho_a, rho_b, M, UID)
    ! UID   : Output file unit ID.
    implicit none
    integer   , intent(in) :: M, UID
-   complex*16, intent(in) :: rho_a(M,M), rho_b(M,M)
+   complex(kind=8), intent(in) :: rho_a(M,M), rho_b(M,M)
 
    rewind(UID)
    call write_sqmatrix(rho_a, M, UID)
@@ -215,7 +215,7 @@ subroutine write_rho_restart_ocs(rho_a, rho_b, M, UID)
    ! UID   : Output file unit ID.
    implicit none
    integer  , intent(in) :: M, UID
-   complex*8, intent(in) :: rho_a(M,M), rho_b(M,M)
+   complex(kind=4), intent(in) :: rho_a(M,M), rho_b(M,M)
 
    rewind(UID)
    call write_sqmatrix(rho_a, M, UID)
