+++ lioamber/liosubs/atmvec_to_orbvec.header.f90
@@ -25,17 +25,17 @@ end subroutine atmvec_to_orbvec_d
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%!
 subroutine atmvec_to_orbvec_c( atmvec, atm_of_orb, orbvec )
    implicit none
-   complex*8 , intent(in)  :: atmvec(:)
+   complex(kind=4) , intent(in)  :: atmvec(:)
    integer   , intent(in)  :: atm_of_orb(:)
-   complex*8 , intent(out) :: orbvec(:)
+   complex(kind=4) , intent(out) :: orbvec(:)
 #  include "atmvec_to_orbvec.proced.f90"
 end subroutine atmvec_to_orbvec_c
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%!
 subroutine atmvec_to_orbvec_z( atmvec, atm_of_orb, orbvec )
    implicit none
-   complex*16, intent(in)  :: atmvec(:)
+   complex(kind=8), intent(in)  :: atmvec(:)
    integer   , intent(in)  :: atm_of_orb(:)
-   complex*16, intent(out) :: orbvec(:)
+   complex(kind=8), intent(out) :: orbvec(:)
 #  include "atmvec_to_orbvec.proced.f90"
 end subroutine atmvec_to_orbvec_z
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%!
