+++ lioamber/liosubs/check_matsize.header.f90
@@ -33,7 +33,7 @@ subroutine check_matsize_c( nsize1, nsize2, matrix, matname, subname )
    implicit none
    integer         , intent(in) :: nsize1
    integer         , intent(in) :: nsize2
-   complex*8       , intent(in) :: matrix(:,:)
+   complex(kind=4)       , intent(in) :: matrix(:,:)
    character(len=*), intent(in) :: matname
    character(len=*), intent(in) :: subname
 #  include "check_matsize.proced.f90"
@@ -43,7 +43,7 @@ subroutine check_matsize_z( nsize1, nsize2, matrix, matname, subname )
    implicit none
    integer         , intent(in) :: nsize1
    integer         , intent(in) :: nsize2
-   complex*16      , intent(in) :: matrix(:,:)
+   complex(kind=8)      , intent(in) :: matrix(:,:)
    character(len=*), intent(in) :: matname
    character(len=*), intent(in) :: subname
 #  include "check_matsize.proced.f90"
