+++ lioamber/liosubs/check_vecsize.header.f90
@@ -29,7 +29,7 @@ end subroutine check_vecsize_d
 subroutine check_vecsize_c( nsize, vector, vecname, subname )
    implicit none
    integer         , intent(in) :: nsize
-   complex*8       , intent(in) :: vector(:)
+   complex(kind=4)       , intent(in) :: vector(:)
    character(len=*), intent(in) :: vecname
    character(len=*), intent(in) :: subname
 #  include "check_vecsize.proced.f90"
@@ -38,7 +38,7 @@ end subroutine check_vecsize_c
 subroutine check_vecsize_z( nsize, vector, vecname, subname )
    implicit none
    integer         , intent(in) :: nsize
-   complex*16      , intent(in) :: vector(:)
+   complex(kind=8)      , intent(in) :: vector(:)
    character(len=*), intent(in) :: vecname
    character(len=*), intent(in) :: subname
 #  include "check_vecsize.proced.f90"
