+++ lioamber/liosubs/read_list.header.f90
@@ -23,14 +23,14 @@ end subroutine read_list_d
 subroutine read_list_c( file_name, list )
    implicit none
    character(len=*), intent(in)  :: file_name
-   complex*8       , intent(out) :: list(:)
+   complex(kind=4)       , intent(out) :: list(:)
 #  include "read_list.proced.f90"
 end subroutine read_list_c
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%!
 subroutine read_list_z( file_name, list )
    implicit none
    character(len=*), intent(in)  :: file_name
-   complex*16      , intent(out) :: list(:)
+   complex(kind=8)      , intent(out) :: list(:)
 #  include "read_list.proced.f90"
 end subroutine read_list_z
 !%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%!
